Separate all types into their own submodule
Fixes #9
